🎨 Rename also types and clean up
