+++ examples/driver/dynamic/CMakeLists.txt
@@ -10,11 +10,10 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
-add_library(qdmi_example_dynamic_driver qdmi_example_dynamic_driver.cpp
-                                        qdmi_example_dynamic_driver.h)
-target_link_libraries(qdmi_example_dynamic_driver PRIVATE qdmi::qdmi)
-target_include_directories(qdmi_example_dynamic_driver
+add_library(qdmi_example_driver qdmi_example_driver.cpp qdmi_example_driver.h)
+target_link_libraries(qdmi_example_driver PRIVATE qdmi::qdmi)
+target_include_directories(qdmi_example_driver
                            PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
-set_target_properties(qdmi_example_dynamic_driver
-                      PROPERTIES POSITION_INDEPENDENT_CODE ON)
-add_library(qdmi::example_dynamic_driver ALIAS qdmi_example_dynamic_driver)
+set_target_properties(qdmi_example_driver PROPERTIES POSITION_INDEPENDENT_CODE
+                                                     ON)
+add_library(qdmi::example_driver ALIAS qdmi_example_driver)
