+++ test/utils/CMakeLists.txt
@@ -2,7 +2,7 @@
 # whether all functions are implemented in the device.
 add_library(qdmi_test_impl SHARED test_impl.cpp test_impl.hpp)
 target_link_libraries(
-  qdmi_test_impl PRIVATE gtest qdmi::qdmi qdmi::example_dynamic_driver
+  qdmi_test_impl PRIVATE gtest qdmi::qdmi qdmi::example_driver
                          qdmi::project_warnings)
 target_include_directories(qdmi_test_impl PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 add_library(qdmi::test_impl ALIAS qdmi_test_impl)
